+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to
 ### Added
 
 - `state` subcommand
+- `current` subcommand
 
 ## [0.1.0] - 2025-07-15